Updated my JS!
I removed the animations for each tile because it is interfering with my CSS transform property on the tiles itself. I also added more spin on the top share button at the top right. I will make additional changes if that random Y axis animation messes with the responsiveness of the website(scrollbar at the bottom).